+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 ### Added
 
 - Add support for templating App CRs in organization namespace.
+- Add `--catalog-namespace` flag to `template app`.
 
 ## [1.60.0] - 2022-01-27